+++ firmware/application/CMakeLists.txt
@@ -113,9 +113,11 @@ set(CSRC
 	${BOARDSRC}
 	${FATFSSRC}
 	firmware_info.c
-	usb_serial.c
+	usb_serial_cdc.c
 	usb_serial_descriptor.c
 	usb_serial_endpoints.c
+	usb_serial_io.c
+	usb_serial_shell.c
 	${HACKRF_PATH}/firmware/common/usb.c
 	${HACKRF_PATH}/firmware/common/usb_queue.c
 	${HACKRF_PATH}/firmware/hackrf_usb/usb_device.c
@@ -200,6 +202,7 @@ set(CPPSRC
 	log_file.cpp
 	metadata_file.cpp
 	portapack.cpp
+	usb_serial_event.cpp
 	usb_serial.cpp
 	qrcodegen.cpp
 	radio.cpp
