+++ firmware/application/event_m0.cpp
@@ -44,10 +44,6 @@ using namespace lpc43xx;
 
 #include "ui_navigation.hpp"
 
-namespace portapack {
-void createShellOnDemand();
-}
-
 extern "C" {
 
 CH_IRQ_HANDLER(M4Core_IRQHandler) {
@@ -117,8 +113,7 @@ void EventDispatcher::run() {
     while (is_running) {
         const auto events = wait();
         dispatch(events);
-
-        portapack::createShellOnDemand();
+        portapack::usb_serial.dispatch();
     }
 }
